requirements.txt
# Core dependencies
websocket-client==1.6.4
aiohttp==3.9.1
numpy==1.26.2
pandas==2.1.4

# Data storage
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic==2.7.4
pydantic-settings==2.6.1
loguru==0.7.2

# Monitoring
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1